Section 3 end
A higher or lower app.
Basically, A random number is generated by the compiler.
The User is asked to guess it.
With clues 'Lower' and 'Higher', he should eventually reach the number.
When he does guess it, ' Correct' shows up.
Also, if he wants another try, he clicks TRY AGAIN.