A Guess the Celebrity App
This app features celebrities and 4 options.
With source as the posh24/celebrities site, the  URLs and Names of all
the celebrities is obtained using Pattern And Matching into Lists.
Once, the content to work with is created at the beginning in the
OnCreate method, the createNewQuestion is invoked to place the correct
answer in any of the four options and wrong in the rest.
In the onClick function, the user gets a toast saying 'Correct'  if the
answer is right, 'Wrong! The correct answer is "so and so"' if wrong.

Also, DownloadImages is done using AsyncTask with Bitmap.

DownloadTask is used to read in the input source code of that site we
used as help.

Drawback - Changes in the site format, affects app working.