Basics Of using JSONObject and JSONArray
Effective way to get information(in this case, weather info) from an
online API source (in this case, Open Weather Map).
JSON is effective as it can use the information from API's directly and
also, the app won't break easily as API's don't change their HTML as
often as websites.